Работа над ошибками 15,04
