+++ hwDraftComments_Модульный_JS/assistants.js
@@ -1,7 +1,3 @@
-import { renderCommentators, сommentators } from "./renderCommentators.js";
-
-const buttonElement = document.getElementById ('add-button');
-
 export const currentDate = (date) => {
  let dateTime = new Date(date);
 
@@ -21,44 +17,6 @@ export function delay(interval = 300) {
       }, interval);
     });
 };
-
-const formElement = document.getElementById ('form');
- formElement.addEventListener('keyup', keyEvent);
-  function keyEvent(e) {
-    if (e.code === 'Enter') {
-     buttonElement.dispatchEvent(new Event('click'));
-    }
-  };
- 
-const removeButton = document.getElementById("deleteComment");
-removeButton.addEventListener("click", () => {
-  сommentators.pop();
-  renderCommentators();
-});
-
-// export const request_retry = async({ url, options, n }) => {
-//   try {
-//     return await fetch(url, options);
-//   } catch (e) {
-//     if (n <= 1) throw e;
-//     await sleep(1000);
-//     return await request_retry({url, options, n: n - 1});
-//   }
-// };
-// request_retry(url, options, 60);
-  
-// const handlePostClick = () => {
-//   //...
-//   postComment(text.value, name.value)
-//     .catch((error) => {
-//       if (error.message === "Ошибка сервера") {
-//         handlePostClick();
-//       }
-//     });
-//   //...
-// };
-
-// addButton.addEventListener("click", handlePostClick);
