+++ hwDraftComments_Модульный_JS/main.js
@@ -29,8 +29,16 @@ function getCom() {
    addForm.classList.remove("hidden");
    preloader.classList.add('preloader-hidden');
 
+  })
+  .catch((error) => {
+    if (error.message === "Сервер упал") {
+      alert("Нет интернета");
+    }
+    if (error.message === "Failed to fetch") {
+      alert("Кажется что-то пошло не так, попробуй позже..");
+    };
   });
-}
+};
 
 getCom();
 
@@ -51,9 +59,19 @@ buttonElement.addEventListener("click", () => {
     nameElement.value = "";
     textElement.value = "";
     return getCom();
-    
-  }).finally(() => loader.textContent = '');
-
+  })
+  .catch((error) => {
+    if (error.message === "Сервер упал") {
+      alert("Нет интернета");
+    }
+    if (error.message === "Вводимые данные слишком короткие") {
+      alert("Имя или текст менее трех символов");
+    }
+    if (error.message === "Failed to fetch") {
+      alert("Кажется что-то пошло не так, попробуй позже..");
+    };
+  })
+  .finally(() => loader.textContent = '');
 });
 
 renderCommentators();
